requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
Pillow>=9.5.0
ptflops>=0.7.2
scikit-learn>=1.3.0
